Refactor tests Moved tests to the correct packages
