+++ src/test/java/brokeculator/frontend/UITest.java
@@ -1,4 +1,4 @@
-package seedu.brokeculator;
+package brokeculator.frontend;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import brokeculator.frontend.UI;
